Move generic Form class to generic form file
