+++ warehouse/accounts/views.py
@@ -74,7 +74,8 @@ from warehouse.email import (
 )
 from warehouse.events.tags import EventTag
 from warehouse.metrics.interfaces import IMetricsService
-from warehouse.oidc.forms.github import DeletePublisherForm, PendingGitHubPublisherForm
+from warehouse.oidc.forms import DeletePublisherForm
+from warehouse.oidc.forms.github import PendingGitHubPublisherForm
 from warehouse.oidc.interfaces import TooManyOIDCRegistrations
 from warehouse.oidc.models import PendingGitHubPublisher, PendingOIDCPublisher
 from warehouse.organizations.interfaces import IOrganizationService
