+++ warehouse/oidc/forms/__init__.py
@@ -9,3 +9,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import wtforms
+
+from warehouse import forms
+from warehouse.i18n import localize as _
+
+
+class DeletePublisherForm(forms.Form):
+    __params__ = ["publisher_id"]
+
+    publisher_id = wtforms.StringField(
+        validators=[
+            wtforms.validators.InputRequired(message=_("Specify a publisher ID")),
+            wtforms.validators.UUID(message=_("Publisher must be specified by ID")),
+        ]
+    )
